pyproject.toml
[tool.poetry]
name = "athena"
version = "0.1.0"
description = ""
authors = ["jeremy.wayland <jeremy.don.wayland@gmail.com>", "stuart.wayland <swayland@ucsc.edu>", "sid.gathrid <sgathrid@yahoo.com>"]
readme = "README.md"
packages = [{ include = "athena" }]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.24.2"
networkx = "^3.0"
black = "^23.1.0"
pandas = "^1.5.3"
jupyter = "^1.0.0"
python-dotenv = "^0.21.1"
pytest = "^7.2.1"
datetime = "^5.0"
scikit-learn = "^1.2.1"
seaborn = "^0.12.2"
python-log-indenter = "^0.9"
numba = "^0.57.1"


urllib3 = "1.26.16"
category-encoders = "^2.6.1"
torch = "^2.0.1"
azure-storage-blob = "^12.17.0"
openpyxl = "^3.1.2"
pympler = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
